HW-0000 Travel agency service
- refactored decorator
